Create component CreateTodoList
