+++ package.json
@@ -62,6 +62,7 @@
     "file-loader": "^6.0.0",
     "formik": "^2.1.4",
     "husky": "^4.2.3",
+    "lodash": "^4.17.15",
     "mini-css-extract-plugin": "^0.9.0",
     "node-sass": "^4.13.1",
     "postcss-loader": "^3.0.0",
@@ -84,7 +85,6 @@
     "history": "^4.10.1",
     "jest": "^25.1.0",
     "jquery": "^3.5.0",
-    "lodash": "^4.17.15",
     "moment": "^2.25.3",
     "popper.js": "^1.16.1",
     "prop-types": "^15.7.2",
